Removed the medaitor list page regex
The regex was causing a failure due to how certain repos are named using "openhim-mediator".
We expect this to be the start of the repo name but some repos have this naming in the middle. Our Regex therefor doesnt return a value we require to display to the user and causes it to break